requirements.txt
prettytable==3.12.0
PyYAML==6.0.2
termcolor==2.5.0
